checking err != nil in gitlab case in service.go file

+++ service.go
@@ -1,10 +1,11 @@
 package main
 
 import (
-	"github.com/urvil38/git-push/utils"
 	"fmt"
 	"os"
 	"path/filepath"
+
+	"github.com/urvil38/git-push/utils"
 
 	"github.com/urvil38/git-push/bitbucket"
 	"github.com/urvil38/git-push/github"
@@ -12,11 +13,11 @@ import (
 	"github.com/urvil38/git-push/types"
 )
 
-func service(service types.Service,repo types.Repo) error {
+func service(service types.Service, repo types.Repo) error {
 	if err := service.Init(); err != nil {
 		return err
 	}
-	
+
 	if err := service.CreateRepo(repo); err != nil {
 		return err
 	}
@@ -28,15 +29,15 @@ func service(service types.Service,repo types.Repo) error {
 	return service.PushRepo()
 }
 
-func invokeService(serviceName string,repo types.Repo) error {
+func invokeService(serviceName string, repo types.Repo) error {
 	switch serviceName {
 	case "GitHub":
-		return service(github.GithubService,repo)
+		return service(github.GithubService, repo)
 	case "BitBucket":
-		return service(bitbucket.BitbucketService,repo)
+		return service(bitbucket.BitbucketService, repo)
 	case "GitLab":
-		err := service(gitlab.GitlabService,repo)
-		if err.Error() == "giterror" {
+		err := service(gitlab.GitlabService, repo)
+		if err != nil && err.Error() == "giterror" {
 			fmt.Printf("%s\n", red("Error: "+"Please check gitlab username or password are correct"))
 			removeFileErr := os.Remove(filepath.Join(utils.GetConfigFolderPath(), "git-push-gitlab"))
 			if removeFileErr != nil {
